requirements.txt
anyio==3.5.0
asgiref==3.5.0
click==8.0.4
fastapi==0.74.1
h11==0.13.0
idna==3.3
pydantic==1.9.0
sniffio==1.2.0
starlette==0.17.1
typing_extensions==4.1.1
uvicorn==0.17.5
joblib==1.1.0
numpy==1.22.2
scipy==1.8.0
sklearn==0.0
threadpoolctl==3.1.0
python-dotenv==0.19.2
SQLAlchemy==1.4.31
mysqlclient==2.1.0